[FIX] fix name of files and classes
