+++ hr_timesheet_default_analytic_account/models/__init__.py
@@ -1,3 +1,1 @@
-from . import account_analytic_account
-from . import hr_employee
-from . import hr_timesheet_sheet
+from . import hr_employee, hr_timesheet_sheet, project_project
